+++ src/main/kotlin/net/fallingangel/jimmercraft/completion/CompletionContributor.kt
@@ -1,0 +1,31 @@
+package net.fallingangel.jimmercraft.completion
+
+import com.intellij.codeInsight.completion.CompletionContributor
+import com.intellij.codeInsight.completion.CompletionParameters
+import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.completion.CompletionType
+import com.intellij.patterns.ElementPattern
+import com.intellij.psi.PsiElement
+
+open class CompletionContributor : CompletionContributor() {
+    /**
+     * 提示指定位置的内容
+     *
+     * @param place 元素位置表达式
+     * @param provider 内容提示
+     */
+    protected fun complete(
+        place: ElementPattern<PsiElement>,
+        provider: (parameters: CompletionParameters, result: CompletionResultSet) -> Unit,
+    ) {
+        extend(
+            CompletionType.BASIC,
+            place,
+            object : CompletionProvider() {
+                override fun completions(parameters: CompletionParameters, result: CompletionResultSet) {
+                    provider(parameters, result)
+                }
+            },
+        )
+    }
+}