+++ src/main/kotlin/net/fallingangel/jimmercraft/facts/Rules.kt
@@ -8,9 +8,9 @@ import net.fallingangel.jimmerdto.lsi.jimmer.JimmerAnnotations
 object Rules : Facet<LName, List<Rule>> {
     override fun get(key: LName): List<Rule> {
         return when (key) {
-            JimmerAnnotations.OneToMany -> listOf(MappedBy(JimmerAnnotations.ManyToOne))
-            JimmerAnnotations.ManyToMany -> listOf(MappedBy(JimmerAnnotations.ManyToMany))
-            JimmerAnnotations.OneToOne -> listOf(MappedBy(JimmerAnnotations.OneToOne))
+            JimmerAnnotations.OneToMany -> listOf(MappedBy.OneToMany)
+            JimmerAnnotations.ManyToMany -> listOf(MappedBy.ManyToMany)
+            JimmerAnnotations.OneToOne -> listOf(MappedBy.OneToOne)
             else -> emptyList()
         }
     }