+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ### Added
 
 - Report invalid property names in `mappedBy` of `@OneToOne`, `@OneToMany` and `@ManyToMany`.
+- Complete valid property names in `mappedBy` of `@OneToOne`, `@OneToMany` and `@ManyToMany`.